+++ src/acceleration/AccelerationGrid_t.cpp
@@ -27,8 +27,8 @@ AccelerationGrid_t::AccelerationGrid_t(Shape_t** items, size_t n_items, size_t m
     bounding_box_ = Box_t(coordinates);
 
     const Vec3f cell_res = (grid_size * std::pow(n_obj_/(grid_size[0]*grid_size[1]*grid_size[2]), 1.0/3.0)).floor()
-                            .max(min_res_)
-                            .min(max_res_) - 1.0;
+                            .max(static_cast<double>(min_res_))
+                            .min(static_cast<double>(max_res_)) - 1.0;
 
     for (unsigned int i = 0; i < 3; ++i) {
         cell_res_[i] = static_cast<size_t>(cell_res[i]+1.0);
@@ -77,8 +77,8 @@ AccelerationGrid_t::AccelerationGrid_t(Shape_t** items, size_t n_items, std::arr
     bounding_box_ = Box_t(coordinates);
 
     const Vec3f cell_res = (grid_size * std::pow(n_obj_/(grid_size[0]*grid_size[1]*grid_size[2]), 1.0/3.0)).floor()
-                            .max(min_res_)
-                            .min(max_res_) - 1.0;
+                            .max(static_cast<double>(min_res_))
+                            .min(static_cast<double>(max_res_)) - 1.0;
 
     for (unsigned int i = 0; i < 3; ++i) {
         cell_res_[i] = static_cast<size_t>(cell_res[i]+1.0);
@@ -126,7 +126,7 @@ AccelerationGrid_t::~AccelerationGrid_t() {
 
 Shape_t* AccelerationGrid_t::intersect(const Ray_t &ray, double &t, std::array<double, 2> &uv) const {
     double tbbox;
-    std::array<int, 3> cellexit {0, 0, 0};
+    std::array<long long, 3> cellexit {0, 0, 0};
     std::array<int, 3> cellstep {0, 0, 0};
     const std::array<unsigned int, 8> map {2, 1, 2, 1, 2, 2, 0, 0};
 
@@ -141,7 +141,7 @@ Shape_t* AccelerationGrid_t::intersect(const Ray_t &ray, double &t, std::array<d
     Vec3f tnext = Vec3f();
 
     const Vec3f raycellorigin = (ray.origin_ + ray.direction_ * tbbox) - bounding_box_.coordinates_[0];
-    const Vec3f cell_res = Vec3f(cell_res_[0], cell_res_[1], cell_res_[2]);
+    const Vec3f cell_res = Vec3f(static_cast<double>(cell_res_[0]), static_cast<double>(cell_res_[1]), static_cast<double>(cell_res_[2]));
     const Vec3f cellcoord = (raycellorigin / cell_size_).floor()
                             .max(0.0)
                             .min(cell_res-1.0);
@@ -156,7 +156,7 @@ Shape_t* AccelerationGrid_t::intersect(const Ray_t &ray, double &t, std::array<d
         else {
             deltat[i] = cell_size_[i] * invdir[i];
             tnext[i] = tbbox + ((cellcoord[i] + 1.0) * cell_size_[i] - raycellorigin[i]) * invdir[i];
-            cellexit[i] = cell_res_[i];
+            cellexit[i] = static_cast<long long>(cell_res_[i]);
             cellstep[i] = 1; 
         }
     }
@@ -195,7 +195,7 @@ void AccelerationGrid_t::add(Shape_t* item) {
     min1.max(0.0);
     max1.max(0.0);
 
-    const Vec3f cell_res = Vec3f(cell_res_[0], cell_res_[1], cell_res_[2]) - 1.0;
+    const Vec3f cell_res = Vec3f(static_cast<double>(cell_res_[0]), static_cast<double>(cell_res_[1]), static_cast<double>(cell_res_[2])) - 1.0;
     min1.min(cell_res);
     max1.min(cell_res);
 
@@ -222,7 +222,7 @@ void AccelerationGrid_t::remove(const Shape_t* item) {
     min1.max(0.0);
     max1.max(0.0);
 
-    const Vec3f cell_res = Vec3f(cell_res_[0], cell_res_[1], cell_res_[2]) - 1.0;
+    const Vec3f cell_res = Vec3f(static_cast<double>(cell_res_[0]), static_cast<double>(cell_res_[1]), static_cast<double>(cell_res_[2])) - 1.0;
     min1.min(cell_res);
     max1.min(cell_res);
